% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{pkg_authors}
\alias{pkg_authors}
\title{Package author and copyright holder/funder}
\usage{
pkg_authors(first = "Matthew", last = "Leonawicz",
  email = "mfleonawicz@alaska.edu", cph_fnd = pkg_cph())
}
\arguments{
\item{first}{author first name.}

\item{last}{author last name.}

\item{email}{author email address}

\item{cph_fnd}{copyright holder and funder.}
}
\value{
a list to be used by \code{use_authors}.
}
\description{
Specify the package author and the copyright holder/funder.
}
\details{
This function is specific to the SNAPverse and is used as a default argument to \code{use_authors}.
}
\examples{
\dontrun{pkg_authors()}
}
