% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{pkg_cph}
\alias{pkg_cph}
\title{Copyright holder}
\usage{
pkg_cph(cph = "Scenarios Network for Alaska and Arctic Planning")
}
\arguments{
\item{cph}{copyright holder.}
}
\value{
copyright holder.
}
\description{
Specify the copyright holder.
}
\details{
This function simply returns its input, the copyright holder string. It is used as a default argument to \code{use_authors}.
}
\examples{
pkg_cph()
}
