% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{pkg_titles}
\alias{pkg_titles}
\title{Get heading information related to SNAPverse R packages}
\usage{
pkg_titles(type)
}
\arguments{
\item{type}{character, \code{"label"}, \code{"title"} or \code{"sub"}.}
}
\value{
a vector of headings.
}
\description{
This function returns a list of headings related to SNAPverse R packages for \code{apputils::app_showcase} display purposes.
}
\details{
Labels are generally shown above an app showcase widget. Titles are generally the top line heading over the image link inside the widget.
Subtitles are for smaller text below.
}
\examples{
pkg_titles("label")
}
