% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{sector_members}
\alias{sector_members}
\title{Get members of a sector}
\usage{
sector_members(pkg)
}
\arguments{
\item{pkg}{character, a sector package. See details.}
}
\value{
a character vector.
}
\description{
Get a list of the collection of member packages that are installed and loaded via a SNAPverse sector package.
}
\details{
The available sector packages are \code{"snapverse"}, \code{"snaplite"}, \code{"snapdata"} and \code{"snapwebs"}.
}
\examples{
sector_members("snapverse")
}
