% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectors.R
\name{sector_update}
\alias{sector_update}
\title{Update SNAPverse sector packages}
\usage{
sector_update(pkg, force = FALSE, quiet = TRUE)
}
\arguments{
\item{pkg}{character, the sector package covering all member packages.}

\item{force}{If \code{TRUE}, force member install even if unchanged.}

\item{quiet}{logical.}
}
\description{
Update a SNAPverse sector package and its member packages.
}
\examples{
\dontrun{
sector_update("snapverse")
}
}
