% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{sv_local_pkgs}
\alias{sv_local_pkgs}
\title{List local SNAPverse packages}
\usage{
sv_local_pkgs(base_path = "../", self = TRUE)
}
\arguments{
\item{base_path}{defaults to the parent directory of the working directory.}

\item{self}{logical, include the currently loaded package (working directory) if in the SNAPverse.}
}
\value{
a character vector.
}
\description{
Generate a list of existing local SNAPverse packages.
}
\details{
This function looks for packages beginning with the expression \code{^snap*} and sharing the same parent
directory as the current package/working directory. If \code{self = TRUE} (default),
the current working directory will be included in the list if applicable.
}
\examples{
sv_local_pkgs()
}
