% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{sv_pkgs}
\alias{sv_pkgs}
\title{Location in the SNAPverse}
\usage{
sv_pkgs()
}
\value{
a data frame.
}
\description{
List the packages in and around the SNAPverse and their section of the verse.
}
\details{
This function returns a 3-column data frame of all SNAPverse package names, their part of the verse,
and whether local git repositories/R projects sharing the same parent directory are found for each.
The list includes SNAPverse satellite packages, which are related to and often used in conjunction with,
but not a strict part of the SNAPverse.

The other packages are considered core packages of the SNAPverse and fall into of three sectors of the verse:
functions, data or apps. The three packages listed as \code{sector} packages load different subsets of SNAPverse packages
for convenience. \code{snapverse} loads functions, data and apps packages. \code{snaplite} loads only the functions packages.
\code{snapdata} loads only the data packages. For most users, the packages most likely needed or of interest
are \code{snapstat}, \code{snapapps} and any of the data packages.
}
\examples{
sv_pkgs
}
