% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapapps.R
\name{use_appdesc}
\alias{use_appdesc}
\title{Create an app DESCRIPTION file}
\usage{
use_appdesc(base_path = ".", title, author = "Matthew Leonawicz",
  url = "https://leonawicz.github.io/", license = "MIT",
  mode = "Showcase", tags)
}
\arguments{
\item{base_path}{output directory.}

\item{title}{app title.}

\item{author}{author name.}

\item{url}{author url.}

\item{license}{license type.}

\item{mode}{Shiny display mode.}

\item{tags}{optional tags. May be a vector. See details.}
}
\description{
Add a DESCRIPTION template to a Shiny app.
}
\details{
This file is used with Shiny app showcase mode.
Since \code{snapmeta} is a developer package used by the SNAPverse author/maintainer,
most arguments to \code{use_appdesc} have contextual defaults that you must otherwise override.
\code{title} and \code{tags} do not have defaults. If missing, they should be updated directly in the generated DESCRIPTION file.
Tags passed in the form \code{tags = c("random variables", "plotmath")} will result in \code{random-variables, plotmath}.
}
\examples{
\dontrun{use_appdesc()}
}
