% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapapps.R
\name{use_appreadme}
\alias{use_appreadme}
\title{Create an app Readme.md file}
\usage{
use_appreadme(base_path = ".")
}
\arguments{
\item{base_path}{output directory.}
}
\description{
Add a Readme.md template to a Shiny app.
}
\details{
This file is used to add a text description below a Shiny app when run in showcase mode.
}
\examples{
\dontrun{use_appreadme()}
}
