% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{use_authors}
\alias{use_authors}
\title{Add author and license information to R package}
\usage{
use_authors(authors = pkg_authors(), cph = pkg_cph())
}
\arguments{
\item{authors}{list as returned by \code{pkg_authors}.}

\item{cph}{copyright holder as returned by \code{pkg_cph}.}
}
\value{
side effect of updating DESCRIPTION and adding MIT license.
}
\description{
Add author and MIT license to DESCRIPTION and add license files to R package.
The defaults for this function cater to the SNAPverse.
}
\examples{
\dontrun{use_authors()}
}
