% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{use_clone_notes}
\alias{use_clone_notes}
\title{Add \code{clone_notes.md} to R package}
\usage{
use_clone_notes()
}
\value{
side effect of creating file.
}
\description{
Add \code{clone_notes.md} to a SNAPverse R package.
}
\details{
The created file is for tracking information unique to a package and helpful to know upon cloning
the repository, e.g., external locations of large data sets, available cached files, or other required files that are not tracked with git.
}
\examples{
\dontrun{use_clone_notes}
}
