% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{use_hex}
\alias{use_hex}
\title{Add \code{hex.R} to R package}
\usage{
use_hex()
}
\value{
side effect of creating file.
}
\description{
Add \code{hex.R} to a SNAPverse R package \code{data-raw} directory.
}
\details{
This script provides a template for package hex sticker icons, a large icon in \code{data-raw} and a small icon in \code{inst}.
The script is not run by \code{snapmeta} since it will likely require customization after it is generated.
}
\examples{
\dontrun{use_hex}
}
