% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{use_these}
\alias{use_these}
\title{Wrapper function for package setup}
\usage{
use_these(pkg = basename(getwd()), authors = pkg_authors(),
  cph = pkg_cph())
}
\arguments{
\item{pkg}{defaults to the name of the working directory.}

\item{authors}{list as generated by \code{pkg_authors}.}

\item{cph}{copyright holder as generated by \code{pkg_cph}.}
}
\value{
side effect of setting up various package files and configurations.
}
\description{
Wrapper function around several package creation/setup functions from the \code{usethis} and \code{snapmeta} packages.
}
\details{
\code{use_these} calls the following functions: \code{use_authors}, \code{use_github_links}, \code{use_clone_notes},
\code{use_cran_comments},\code{use_data_raw}, \code{use_news_md}, \code{use_testthat}, \code{use_vignette},
\code{use_readme_rmd}, \code{use_revdep}, \code{use_lintr}, \code{use_appveyor}, \code{use_travis} and \code{use_coverage}.

Authors, clone notes and \code{lintr} use are by \code{snapmeta} functions. The others are \code{usethis} package functions,
except for \code{use_vignettes}, which still relies on \code{devtools}.
\code{pkgdown} for R package website building is also initialized, using a \code{pkgdown} directory in the package root
directory containing template \code{_pkgdown.yml} and \code{extra.css} files relevant to the SNAPverse.
The \code{docs} directory is used for website files and should be specified likewise in the repository settings on GitHub.

\code{pkgdown::init_site} is also called.
\code{.Rbuildignore} is also updated. Make sure to remove the absolute path created by \code{pkgdown::init_site}.
Only retain the relative path to the \code{docs} directory.
Finally, it adds \code{snapmeta} under the \code{DESCRIPTION} file \code{Imports} field for SNAPverse sector packages.
If needed for any other packages, it must be added manually.
}
\examples{
\dontrun{use_these()}
}
