% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectors.R
\name{verse_attach}
\alias{verse_attach}
\alias{is_attached}
\title{Attach SNAPverse member packages}
\usage{
verse_attach(core)

is_attached(x)
}
\arguments{
\item{core}{character, vector of packages that defines part of the SNAPverse.}

\item{x}{character, package name.}
}
\description{
Attach SNAPverse member packages for different sectors of the verse.
There are three sector packages: \code{snapverse}, \code{snaplite} and \code{snapdata}.
Each passes a list of packages via \code{core} to \code{verse_attach}.
See the respective packages for details.

Checks if a package is attached.
}
\details{
Verse attachment code snippets are borrowed and modified from tidyverse: https://github.com/tidyverse
but are slightly more general to handle different SNAPverse package subsets (sectors of verse).
The functionality in \code{snapmeta} only loads sector packages tidily; it does not check for conflicts.
}
