% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{sv_pkgs}
\alias{sv_pkgs}
\title{Location in the SNAPverse}
\usage{
sv_pkgs()
}
\value{
a data frame.
}
\description{
List the packages in and around the SNAPverse and their section of the verse.
}
\details{
This function returns a 3-column data frame of all SNAPverse package names, their part of the verse,
and whether local git repositories/R projects sharing the same parent directory are found for each.
It includes SNAPverse satellite packages.
}
\examples{
sv_pkgs
}
