% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{use_lintr}
\alias{use_lintr}
\title{Add `code{lintr} to R package}
\usage{
use_lintr(base_path = ".")
}
\arguments{
\item{base_path}{package root directory.}
}
\value{
side effect of adding \code{lintr} scaffolding to package.
}
\description{
Add scaffolding for \code{lintr} package usage to R package.
}
\details{
This function creates a template \code{.lintr} file for SNAPverse packages inside \code{inst} and
adds a symbolic link to this file in the package root directory.
It also adds both files to \code{.Rbuildignore}.
}
\examples{
\dontrun{use_lintr()}
}
